% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{check_unique_identifiers}
\alias{check_unique_identifiers}
\title{Identify columns that could serve as a unique key}
\usage{
check_unique_identifiers(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A character vector of candidate key columns, possibly empty.
}
\description{
Identify columns that could serve as a unique key
}
