% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagrade-package.R
\docType{package}
\name{datagrade-package}
\alias{datagrade}
\alias{datagrade-package}
\title{datagrade: scored data quality assessment for open data sets}
\description{
One call, [dg_assess()], profiles a data set and scores it on the ISO/IEC
25024 dimensions of completeness, accuracy, consistency and timeliness. It
returns a structured `dg_report` object with tables and figures rather than
console text alone.
}
\details{
The scoring framework implements the methodology of Demirhan, O. H. (2024),
*Assessing Data Quality Management Issues in Open-Sourced Data Sets with
Statistical Methodology*, MSc dissertation, University of Greenwich. Where the
package deviates from the dissertation, the deviation is recorded in
`NEWS.md` and in the "Deviations from the dissertation" section of the README.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/halukdemirhan/datagrade}
  \item Report bugs at \url{https://github.com/halukdemirhan/datagrade/issues}
}

}
\author{
\strong{Maintainer}: Omer Haluk Demirhan \email{hlk01dn@gmail.com}

Authors:
\itemize{
  \item Omer Haluk Demirhan \email{hlk01dn@gmail.com}
}

}
\keyword{internal}
