% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{print.dg_report}
\alias{print.dg_report}
\title{Print a quality report}
\usage{
\method{print}{dg_report}(x, ...)
}
\arguments{
\item{x}{A `dg_report`.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print a quality report
}
