% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{summary.dg_report}
\alias{summary.dg_report}
\title{Tabular summary of a quality report}
\usage{
\method{summary}{dg_report}(object, ...)
}
\arguments{
\item{object}{A `dg_report`.}

\item{...}{Unused.}
}
\value{
A list of data frames: `scores`, `columns`, `redundant_pairs` and
  `distribution`.
}
\description{
Tabular summary of a quality report
}
